+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Add tests for ProviderEditRoute. (UIEH-1186)
 * Fix 3 GET requests instead of 1 are made when searching Providers/Packages/Titles. (MODKBEKBJ-619)
 * Fix datepicker date format for Custom coverage dates to match current user locale when edit resource. (UIEH-1227)
+* Add tests for useMultiColumnListSort. (UIEH-1183)
 
 ## [7.0.1] (https://github.com/folio-org/ui-eholdings/tree/v7.0.1) (2021-10-19)